// src/Actors/actor.model.d.ts
export interface actorDTO {
  id: number;
  name: string;
  dateOfBirth: Date;
  picture: string;
  biography: string;
}
export interface actorCreationDTO {
  name: string;
  dateOfBirth?: Date;
  picture?: File;
  pictureURL?: string;
  biography?: string;
}
export interface actorMovieDTO {
  id: number;
  name: string;
  character: string;
  picture: string;
}